Switch to bf16-mixed precision for maximum Blackwell throughput
BF16 has float32 range (no overflow risk) with native Tensor Core support.
Combined with set_float32_matmul_precision('medium') covers all matmul paths.

+++ gen_model/train_unconditional.py
@@ -109,7 +109,7 @@ def main():
         max_steps=args.max_steps,
         accelerator='auto', devices='auto',
         callbacks=[checkpoint_cb],
-        precision='16-mixed' if torch.cuda.is_available() else 32,
+        precision='bf16-mixed' if torch.cuda.is_available() else 32,
         log_every_n_steps=500,
         enable_progress_bar=False,
         gradient_clip_val=args.grad_clip,
